Reorganize typeit.schema
