+++ .travis.yml
@@ -15,7 +15,7 @@ python:
 
 # --source specifies what packages to cover, you probably want to use that option
 script:
-    # - mypy --config-file setup.cfg --package typeit
+    # - make typecheck
     - pytest --cov=typeit
 
 after_success: coveralls
