+++ Makefile
@@ -1,0 +1,7 @@
+PROJECT=typeit
+
+test:
+	pytest -s tests/
+
+typecheck:
+	mypy --config-file setup.cfg --package $(PROJECT)
