+++ typeit/__init__.py
@@ -1,5 +1,6 @@
 from .parser import type_constructor
 from .codegen import typeit
 from . import flags
+from .schema.errors import Invalid
 
-__all__ = ['type_constructor', 'typeit', 'flags']
+__all__ = ['type_constructor', 'typeit', 'flags', 'Invalid']
