+++ typeit/cli/run.py
@@ -2,11 +2,12 @@ import argparse
 import json
 import sys
 from pathlib import Path
+from typing import Dict
 
 from .. import codegen as cg
 
 
-def setup(subparsers):
+def setup(subparsers: argparse._SubParsersAction) -> argparse.ArgumentParser:
     sub = subparsers.add_parser('gen', help='Generate hints for structured data (JSON, YAML)')
     sub.add_argument('-s', '--source', help="Path to structured data (JSON, YAML). "
                                             "If not specified, then the data will be read from stdin.")
@@ -14,24 +15,24 @@ def setup(subparsers):
     return sub
 
 
-def main(args: argparse.Namespace):
+def main(args: argparse.Namespace) -> None:
     """ $ typeit gen <source> <target>
     """
     try:
         with Path(args.source).open('r') as f:
-            struct = _read_data(f)
+            dict_struct = _read_data(f)
     except TypeError:
         # source is None, read from stdin
-        struct = _read_data(sys.stdin)
+        dict_struct = _read_data(sys.stdin)
 
-    struct, overrides = cg.typeit(struct)
+    struct, overrides = cg.typeit(dict_struct)
     python_src, __ = cg.codegen_py(struct, overrides)
     sys.stdout.write(python_src)
     sys.stdout.write('\n')
     sys.exit(0)
 
 
-def _read_data(fd):
+def _read_data(fd) -> Dict:
     buf = fd.read()  # because stdin does not support seek
     try:
         struct = json.loads(buf)
@@ -45,5 +46,5 @@ def _read_data(fd):
                 "dependency, or use the `third_party` extra tag with typeit:\n\n"
                 "$ pip install typeit[third_party]"
             )
-        struct = yaml.load(buf, Loader=yaml.FullLoader)
+        struct = yaml.full_load(buf)
     return struct
