+++ typeit/schema/errors.py
@@ -1,0 +1,5 @@
+import colander
+
+# Alias for typeit.Invalid, so users don't have to import two
+# packages
+Invalid = colander.Invalid
