+++ typeit/schema/nodes.py
@@ -7,3 +7,17 @@ class SchemaNode(col.SchemaNode):
     """
     def __repr__(self) -> str:
         return f'SchemaNode({self.typ})'
+
+
+class SetSchema(col.SequenceSchema):
+    def __init__(self, *args, frozen=False, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.frozen = frozen
+
+    def deserialize(self, *args, **kwargs):
+        r = super().deserialize(*args, **kwargs)
+        if r in (col.null, None):
+            return r
+        if self.frozen:
+            return frozenset(r)
+        return set(r)
