+++ typeit/utils.py
@@ -1,7 +1,7 @@
-from typing import Iterator, NamedTuple, Any, Mapping, Union, Sequence, Optional
+from typing import Iterator, NamedTuple, Any, Mapping, Union, Optional
 import re
 import keyword
-import colander
+from .schema.errors import Invalid
 from . import interface as iface
 
 
@@ -25,14 +25,14 @@ def normalize_name(name: str,
     return being_normalized
 
 
-def iter_invalid_data(error: colander.Invalid,
+def iter_invalid_data(error: Invalid,
                       data: Mapping[str, Any]) -> Iterator[InvalidData]:
     """ A helper function to iterate over data samples that
     caused an exception at object construction. Use it like this:
 
     >>> try:
     >>>     obj = mk_obj(data)
-    >>> except colander.Invalid as e:
+    >>> except Invalid as e:
     >>>     # iterate over a sequence of InvalidData
     >>>     for e in iter_invalid_data(e, data):
     >>>         ...
